# openapi.yaml
openapi: 3.1.0
info:
  title: Galaxy of Consequence RPG Backend API
  description: A comprehensive Star Wars RPG backend with NVIDIA Nemotron integration, persistent world simulation, reactive NPCs, and immersive Force mechanics following the Final Rule.
  version: 1.0.0
  contact:
    name: Galaxy of Consequence API
    email: api@galaxyofconsequence.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.galaxyofconsequence.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /create_session:
    post:
      tags:
        - Session
      summary: Create multiplayer session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_name
                - created_by
              properties:
                session_name:
                  type: string
                created_by:
                  type: string
                galaxy_state:
                  type: object
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  session_id:
                    type: string
                  session_name:
                    type: string
                  created_by:
                    type: string
                  galaxy_state:
                    type: object
  /get_session_state:
    get:
      tags:
        - Session
      summary: Get session state
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session state retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionState'

  # Force System
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use 'Abracadabra' as the Bearer token
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /authenticate endpoint
  schemas:
...
    SessionState:
      type: object
      properties:
        session_id:
          type: string
        session_name:
          type: string
        galaxy_state:
          type: object
        active_players:
          type: array
          items:
            type: string
        current_galactic_year:
          type: integer
        major_events:
          type: array
        faction_war_status:
          type: object
        force_nexus_events:
          type: array
        threat_escalation_level:
          type: integer
        last_faction_tick:
          type: string
        updated_at:
          type: string
...